refactor: move original src and test directories to allow initial tutorial state
