+++ test/a11y-input.test.js
@@ -1,9 +1,7 @@
 /* eslint-disable no-unused-expressions */
 import { html, fixture, expect } from '@open-wc/testing';
-import sinon from 'sinon';
 
 import '../src/a11y-input.js';
-import { isMainThread } from 'worker_threads';
 
 /**
  * @typedef {import('../src/a11y-input.js').A11yInput} A11yInput
